fix #3: fix test and docs
