+++ API-Server/src/test/java/swm/hkcc/LGTM/LgtmApplicationTest.java
@@ -1,0 +1,12 @@
+package swm.hkcc.LGTM;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LgtmApplicationTest {
+
+    @Test
+    void main() {
+    }
+}
